Attempt to reorganize state management some more
